gcf-piracy-feed/requirements.txt
requests==2.32.3
google-cloud-storage==2.16.0
